First full clean test run of `build_award_network`.
Added some `roxygen` style headers to some of the functions for my own reference.  I can go in and do some better commenting.

Removed the properties from `cypherToList`, instead I am passing the values explicitly in the `mergeRel()` and `mergeNode()` functions.

`mergeRel()` now renders both the `MATCH` and the `MERGE` statements dynamically.  There's a bug in the big run somewhere, but it works about 30% of the time.