Added parallelization.  This required some changes to the code, but the core functionality is implemented using `multidplyr`.
There are some other changes, in particular, I moved most functions into their own files, so the `R` folder is now populated with `collapse_list`, `mergeNode` and `mergeRel`.

The timing loop in `build_award_network` is going to be dropped, it's just for the Rmd in development to describe the process.