I shuffled the file inputs so that we're not always running from the earliest to the most recent.  Also wrapped the cluster call in a `try()` function so that an error won't break the system.
`parse_award` is not its own file, with the connection to the database made within each function call.

Made a change to `mergeNode` so that any empty nodes get dropped.  I actually think this should be a `grep()` command, but will leave it for now.  This should get caught by the `try()` in `parse_award()`, so I can revisit these elements.